package organization change, controller package, creation of finish layout, managing the last intent, sending information between all intents
