+++ Swoosh/app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".WelcomeActivity">
+        <activity android:name=".Controller.WelcomeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -21,9 +21,10 @@
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
 
-        <activity android:name=".LeagueActivity" />
-        <activity android:name=".BaseActivity" />
-        <activity android:name=".SkillActivity"></activity>
+        <activity android:name=".Controller.LeagueActivity" />
+        <activity android:name=".Controller.BaseActivity" />
+        <activity android:name=".Controller.SkillActivity" />
+        <activity android:name=".Controller.FinishActivity"></activity>
     </application>
 
 </manifest>
