+++ Swoosh/app/src/main/java/com/vinicius/swoosh/Controller/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.vinicius.swoosh
+package com.vinicius.swoosh.Controller
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
