+++ Swoosh/app/src/main/java/com/vinicius/swoosh/Controller/LeagueActivity.kt
@@ -1,9 +1,11 @@
-package com.vinicius.swoosh
+package com.vinicius.swoosh.Controller
 
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import android.widget.Toast
+import com.vinicius.swoosh.Utilities.EXTRA_LEAGUE
+import com.vinicius.swoosh.R
 import kotlinx.android.synthetic.main.activity_league.*
 
 class LeagueActivity : BaseActivity() {
