+++ Swoosh/app/src/main/java/com/vinicius/swoosh/Controller/WelcomeActivity.kt
@@ -1,8 +1,8 @@
-package com.vinicius.swoosh
+package com.vinicius.swoosh.Controller
 
 import android.content.Intent
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import com.vinicius.swoosh.R
 import kotlinx.android.synthetic.main.activity_welcome.*
 
 class WelcomeActivity : BaseActivity() {
