--- Swoosh/app/src/main/java/com/vinicius/swoosh/SkillActivity.kt
@@ -1,18 +1,0 @@
-package com.vinicius.swoosh
-
-import android.content.Intent
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-
-class SkillActivity : BaseActivity() {
-
-    var league = ""
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_skill)
-        league = intent.getStringExtra(EXTRA_LEAGUE)
-        Log.i("info", league)
-    }
-}
