+++ Swoosh/app/src/main/java/com/vinicius/swoosh/Utilities/constants.kt
@@ -1,0 +1,7 @@
+package com.vinicius.swoosh.Utilities
+
+/**
+ * Created by vinicius on 12/22/17.
+ */
+const val EXTRA_LEAGUE = "league"
+const val EXTRA_SKILL = "skill"
