--- Swoosh/app/src/main/java/com/vinicius/swoosh/constants.kt
@@ -1,6 +1,0 @@
-package com.vinicius.swoosh
-
-/**
- * Created by vinicius on 12/22/17.
- */
-const val EXTRA_LEAGUE = "league"
