+++ Swoosh/app/src/main/res/layout/activity_league.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.vinicius.swoosh.LeagueActivity">
+    tools:context="com.vinicius.swoosh.Controller.LeagueActivity">
 
     <ImageView
         android:id="@+id/imageView3"
