+++ Swoosh/app/src/main/res/layout/activity_welcome.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.vinicius.swoosh.WelcomeActivity">
+    tools:context="com.vinicius.swoosh.Controller.WelcomeActivity">
 
     <ImageView
         android:id="@+id/imageView"
