+++ Swoosh/app/src/main/res/values/strings.xml
@@ -15,4 +15,5 @@
     <string name="beginner">Beginner</string>
     <string name="baller">Baller</string>
     <string name="finish">Finish</string>
+    <string name="search_league">Looking for a co-ed beginner league near you...</string>
 </resources>
